Added /create-event path to Navbar, moved CreateEventForm.jsx to src/pages
